face_recognition/requirements.txt
opencv-python
face_recognition
pymongo
python-dotenv
numpy
requests
flask
flask-cors
